+++ packages/react/src/index.ts
@@ -11,6 +11,7 @@ export * from "./save";
 export * from "./achievements";
 export * from "./i18n";
 export * from "./accessibility";
+export * from "./radialMenu";
 export * from "./components";
 export * from "./hud";
 export * from "./bars";